Add retro code

+++ src/routes/room/[id]/components/retrospective/CommentSection.svelte
@@ -1,15 +1,17 @@
 <script lang="ts">
 	import { Input } from '$lib/components/ui/input';
 	import { Button } from '$lib/components/ui/button';
-	import type { RetroItem } from '$lib/scrum/types/retrospective';
+	import type { RetroItem } from '$lib/scrum/types/retro_types';
 	import X from 'lucide-svelte/icons/x';
 
 	let {
 		retroItem,
+		participantId,
 		onAddComment,
 		onDeleteComment
 	}: {
 		retroItem: RetroItem;
+		participantId: string;
 		onAddComment: (content: string) => void;
 		onDeleteComment: (id: string) => void;
 	} = $props();
@@ -28,23 +30,30 @@
 	<h4 class="text-sm font-semibold">Comments</h4>
 	{#each retroItem.comments as comment}
 		<div class="flex items-center rounded-md bg-muted p-2 text-sm">
-			<!--			<p class="font-medium">{comment.author}</p>-->
 			<div>
 				<p>{comment.content}</p>
 			</div>
-			<div class="ml-auto">
-				<Button
-					variant="ghost"
-					size="sm"
-					on:click={() => onDeleteComment(comment.id)}
-					class="h-8 px-2 text-destructive"
-				>
-					<X class="h-4 w-4" />
-				</Button>
-			</div>
+			{#if participantId === comment.author}
+				<div class="ml-auto">
+					<Button
+						variant="ghost"
+						size="sm"
+						on:click={() => onDeleteComment(comment.id)}
+						class="h-8 px-2 text-destructive"
+					>
+						<X class="h-4 w-4" />
+					</Button>
+				</div>
+			{/if}
 		</div>
 	{/each}
-	<form class="flex space-x-2" onsubmit={handleAddComment}>
+	<form
+		class="flex space-x-2"
+		onsubmit={(event) => {
+			event.preventDefault();
+			handleAddComment();
+		}}
+	>
 		<Input placeholder="Add a comment" bind:value={newComment} class="text-sm" />
 		<Button size="sm" on:click={handleAddComment}>Add</Button>
 	</form>
